typeScript implimented
