+++ src/App.tsx
@@ -1,0 +1,28 @@
+// import { RouterProvider } from 'react-router-dom';
+// import './App.css';
+// import router from './Router/Router';
+
+// function App() {
+//   return (
+//     <div>
+//       <RouterProvider router={router}></RouterProvider>
+//     </div>
+//   );
+// }
+
+// export default App;
+
+import React from 'react';
+import { RouterProvider } from 'react-router-dom';
+import './App.css';
+import router from './Router/Router';
+
+function App() {
+  return (
+    <div className="App">
+      <RouterProvider router={router}></RouterProvider>
+    </div>
+  );
+}
+
+export default App;
